Create MappingsRepository and refactor use cases - refactor GetMappingByOwnerUseCase - refactor SaveMappingUseCase
