+++ src/domain/common/factories/RepositoryFactory.ts
@@ -11,6 +11,7 @@ import { DataSource } from "../../instance/entities/DataSource";
 import { Instance } from "../../instance/entities/Instance";
 import { InstanceFileRepositoryConstructor } from "../../instance/repositories/InstanceFileRepository";
 import { InstanceRepositoryConstructor } from "../../instance/repositories/InstanceRepository";
+import { MappingRepositoryConstructor } from "../../mapping/repositories/MappingRepository";
 import { MetadataRepository, MetadataRepositoryConstructor } from "../../metadata/repositories/MetadataRepository";
 import { MigrationsRepositoryConstructor } from "../../migrations/repositories/MigrationsRepository";
 import { GitHubRepositoryConstructor } from "../../packages/repositories/GitHubRepository";
@@ -160,6 +161,13 @@ export class RepositoryFactory {
         const config = this.configRepository(instance);
         return this.get<MigrationsRepositoryConstructor>(Repositories.MigrationsRepository, [config, instance]);
     }
+
+    @cache()
+    public mappingRepository(instance: Instance) {
+        const config = this.configRepository(instance);
+
+        return this.get<MappingRepositoryConstructor>(Repositories.MappingRepository, [config]);
+    }
 }
 
 type RepositoryKeys = typeof Repositories[keyof typeof Repositories];
@@ -184,4 +192,5 @@ export const Repositories = {
     MigrationsRepository: "migrationsRepository",
     TEIsRepository: "teisRepository",
     UserRepository: "userRepository",
+    MappingRepository: "mappingRepository",
 } as const;
